+++ docs/whats-new.rst
@@ -3,6 +3,28 @@
 What's New
 ==========
 
+.. _whats-new.0.1.3:
+
+v0.1.3 (unreleased)
+-------------------
+
+Enhancements
+~~~~~~~~~~~~
+
+- Use ``dask.bag`` coupled with ``dask.distributed`` rather than
+  ``multiprocess`` to parallelize computations (closes :issue:`169` 
+  via :pull:`172`).  This enables the optional use of an external
+  ``distributed.Client`` to leverage computational resources across
+  multiple nodes of a cluster. By `Spencer Clark <https://github.com/spencerkclark>`_.
+
+Dependencies
+~~~~~~~~~~~~
+
+- ``multiprocess`` is no longer required for submitting ``aospy`` calculations
+  in parallel (see discussion in :issue:`169` and pull request :pull:`172`).
+- ``aospy`` now requires an installation of ``dask`` with version greater than or equal
+  to 0.14 (see discussion in pull request :pull:`172`).
+
 .. _whats-new.0.1.2:
 
 v0.1.2 (30 March 2017)